Reorganise UI library
