+++ libs/ui/src/index.ts
@@ -1,10 +1,9 @@
-export * from './lib/CreateBudgetInput/CreateBudgetInput';
-export * from './lib/BudgetMenu/BudgetMenu';
-export * from './lib/CategoryBar/CategoryBar';
-export * from './lib/NavBar/NavBar';
-export * from './lib/Menu/Menu';
-export * from './lib/Menu/MenuItem';
+export * from './lib/Budgets/CreateBudgetInput/CreateBudgetInput';
+export * from './lib/Budgets/BudgetMenu/BudgetMenu';
+export * from './lib/Merchants/MerchantMenu/MerchantMenu';
+export * from './lib/Navigation/NavBar/NavBar';
+export * from './lib/Navigation/Menu/Menu';
 export * from './lib/SwitchButton/SwitchButton'
 export * from './theme/DefaultTheme';
-export * from './lib/MonthlyGraph/MonthlyGraph'
-export * from './lib/BudgetGraph/BudgetGraph'
+export * from './lib/Charts/MonthlyGraph/MonthlyGraph'
+export * from './lib/Charts/BudgetGraph/BudgetGraph'
