+++ libs/ui/src/lib/Charts/MonthlyGraph/MonthlyGraph.stories.tsx
