+++ libs/ui/src/lib/Charts/MonthlyGraph/MonthlyGraph.tsx
