+++ libs/ui/src/lib/Merchants/MerchantMenu/MerchantMenu.stories.tsx
@@ -1,24 +1,26 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { CategoryBar } from './CategoryBar';
+import { MerchantMenu } from './MerchantMenu';
 
 export default {
-  component: CategoryBar,
+  component: MerchantMenu,
   title: 'CategoryBar',
-} as ComponentMeta<typeof CategoryBar>;
+} as ComponentMeta<typeof MerchantMenu>;
 
-const Template: ComponentStory<typeof CategoryBar> = (args) => (
-  <CategoryBar {...args} />
+const Template: ComponentStory<typeof MerchantMenu> = (args) => (
+  <MerchantMenu {...args} />
 );
 
 export const Primary = Template.bind({});
 Primary.args = {
   merchants: [
     {
+      id: 1,
       name: "Bunnings",
       up_category: "home stuff",
-      category: null
+      category: ""
     },
     {
+      id: 2,
       name: "Woolies",
       up_category: "groceries",
       category: "food"
