--- libs/ui/src/lib/NavBar/NavBar.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import NavBar from './NavBar';
-
-describe('NavBar', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<NavBar />);
-    expect(baseElement).toBeTruthy();
-  });
-});
